moved code to mc_sampler.py
